fix:update ApplicationTest and add TestDefaultConfig
